Replaced dropdown Assigned vehicle to crew by a combobox in vehicles tab, and added it to Add Vehicle modal
