+++ components/gang/add-vehicle.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useMemo } from 'react';
+import React, { useState, useEffect } from 'react';
 import { Input } from '../ui/input';
 import Modal from '@/components/ui/modal';
 import { toast } from 'sonner';
@@ -12,11 +12,7 @@ import { addGangVehicle } from '@/app/actions/add-gang-vehicle';
 import { assignVehicleToFighter } from '@/app/actions/assign-vehicle-to-fighter';
 import { getAllowedLocomotionOptions } from '@/utils/vehicle-locomotion';
 import { UserPermissions } from '@/types/user-permissions';
-import { IoSkull } from 'react-icons/io5';
-import { MdChair } from 'react-icons/md';
-import { GiCrossedChains, GiHandcuffs } from 'react-icons/gi';
-import { TbMeatOff } from 'react-icons/tb';
-import { FaMedkit } from 'react-icons/fa';
+import { useCrewFighterOptions } from '@/utils/crew-fighter-combobox-options';
 
 interface VehicleType {
   id: string;
@@ -78,44 +74,7 @@ export default function AddVehicle({
     ? getAllowedLocomotionOptions(selectedVehicleType.vehicle_type)
     : [];
 
-  const crewFighters = fighters.filter(fighter =>
-    fighter.fighter_class === 'Crew' &&
-    (!fighter.vehicles || fighter.vehicles.length === 0)
-  );
-
-  const crewFighterOptions = useMemo(() => {
-    return [...crewFighters]
-      .sort((a, b) => {
-        if (!positioning) return 0;
-        const indexA = Object.entries(positioning).find(([, id]) => id === a.id)?.[0];
-        const indexB = Object.entries(positioning).find(([, id]) => id === b.id)?.[0];
-        const posA = indexA !== undefined ? parseInt(indexA) : Infinity;
-        const posB = indexB !== undefined ? parseInt(indexB) : Infinity;
-        return posA - posB;
-      })
-      .map((f) => {
-        const statusIcons = [];
-        if (f.killed) statusIcons.push(<IoSkull className="text-gray-400 w-4 h-4" key="killed" />);
-        if (f.retired) statusIcons.push(<MdChair className="text-muted-foreground w-4 h-4" key="retired" />);
-        if (f.enslaved) statusIcons.push(<GiCrossedChains className="text-sky-200 w-4 h-4" key="enslaved" />);
-        if (f.starved) statusIcons.push(<TbMeatOff className="text-red-500 w-4 h-4" key="starved" />);
-        if (f.recovery) statusIcons.push(<FaMedkit className="text-blue-500 w-4 h-4" key="recovery" />);
-        if (f.captured) statusIcons.push(<GiHandcuffs className="text-red-600 w-4 h-4" key="captured" />);
-
-        const displayText = `${f.fighter_name} - ${f.fighter_type}${f.xp !== undefined ? ` (${f.xp} XP)` : ''}`;
-
-        return {
-          value: f.id,
-          displayValue: displayText,
-          label: (
-            <span className="flex items-center gap-1">
-              <span>{displayText}</span>
-              {statusIcons.length > 0 && <span className="flex items-center gap-0.5">{statusIcons}</span>}
-            </span>
-          ),
-        };
-      });
-  }, [crewFighters, positioning]);
+  const crewFighterOptions = useCrewFighterOptions(fighters, positioning);
 
   // Fetch vehicle types when component mounts
   useEffect(() => {
@@ -246,30 +205,10 @@ export default function AddVehicle({
         if (assignResult.success) {
           const crewFighter = fighters.find(f => f.id === selectedCrewId);
           if (crewFighter && onFighterUpdate) {
-            const assignedVehicle = {
-              id: newVehicle.id,
-              created_at: newVehicle.created_at,
-              vehicle_name: newVehicle.vehicle_name,
-              vehicle_type_id: newVehicle.vehicle_type_id,
-              vehicle_type: newVehicle.vehicle_type,
-              cost: newVehicle.cost,
-              movement: newVehicle.movement,
-              front: newVehicle.front,
-              side: newVehicle.side,
-              rear: newVehicle.rear,
-              hull_points: newVehicle.hull_points,
-              handling: newVehicle.handling,
-              save: newVehicle.save,
-              body_slots: newVehicle.body_slots,
-              body_slots_occupied: newVehicle.body_slots_occupied,
-              drive_slots: newVehicle.drive_slots,
-              drive_slots_occupied: newVehicle.drive_slots_occupied,
-              engine_slots: newVehicle.engine_slots,
-              engine_slots_occupied: newVehicle.engine_slots_occupied,
-              special_rules: newVehicle.special_rules || [],
-              equipment: newVehicle.equipment || [],
-              effects: {},
-            };
+            // Spread newVehicle, strip fields that don't belong on Vehicle, and
+            // initialise effects to an empty map so the types align.
+            const { gang_id: _g, fighter_id: _f, payment_cost: _p, effects: _e, ...vehicleBase } = newVehicle;
+            const assignedVehicle = { ...vehicleBase, effects: {} };
             onFighterUpdate({
               ...crewFighter,
               vehicles: [assignedVehicle],
