+++ components/gang/vehicles-tab.tsx
@@ -13,15 +13,12 @@ import { deleteVehicle } from '@/app/actions/delete-vehicle';
 import { sellVehicle } from '@/app/actions/sell-vehicle';
 import { UserPermissions } from '@/types/user-permissions';
 import { LuSquarePen } from 'react-icons/lu';
-import { MdCurrencyExchange, MdChair } from 'react-icons/md';
+import { MdCurrencyExchange } from 'react-icons/md';
 import { unassignVehicle } from '@/app/actions/unassign-vehicle';
 import { HiUserRemove } from "react-icons/hi";
 import VehicleEdit from '@/components/gang/vehicle-edit';
 import { Combobox } from '@/components/ui/combobox';
-import { IoSkull } from 'react-icons/io5';
-import { GiCrossedChains, GiHandcuffs } from 'react-icons/gi';
-import { TbMeatOff } from 'react-icons/tb';
-import { FaMedkit } from 'react-icons/fa';
+import { useCrewFighterOptions } from '@/utils/crew-fighter-combobox-options';
 
 interface GangVehiclesProps {
   vehicles: VehicleProps[];
@@ -166,45 +163,7 @@ export default function GangVehicles({
     }
   };
 
-  // Filter for only Crew fighters who don't have vehicles assigned
-  const crewFighters = fighters.filter(fighter => 
-    fighter.fighter_class === 'Crew' && 
-    (!fighter.vehicles || fighter.vehicles.length === 0)
-  );
-
-  const crewFighterOptions = useMemo(() => {
-    return [...crewFighters]
-      .sort((a, b) => {
-        if (!positioning) return 0;
-        const indexA = Object.entries(positioning).find(([, id]) => id === a.id)?.[0];
-        const indexB = Object.entries(positioning).find(([, id]) => id === b.id)?.[0];
-        const posA = indexA !== undefined ? parseInt(indexA) : Infinity;
-        const posB = indexB !== undefined ? parseInt(indexB) : Infinity;
-        return posA - posB;
-      })
-      .map((f) => {
-        const statusIcons = [];
-        if (f.killed) statusIcons.push(<IoSkull className="text-gray-400 w-4 h-4" key="killed" />);
-        if (f.retired) statusIcons.push(<MdChair className="text-muted-foreground w-4 h-4" key="retired" />);
-        if (f.enslaved) statusIcons.push(<GiCrossedChains className="text-sky-200 w-4 h-4" key="enslaved" />);
-        if (f.starved) statusIcons.push(<TbMeatOff className="text-red-500 w-4 h-4" key="starved" />);
-        if (f.recovery) statusIcons.push(<FaMedkit className="text-blue-500 w-4 h-4" key="recovery" />);
-        if (f.captured) statusIcons.push(<GiHandcuffs className="text-red-600 w-4 h-4" key="captured" />);
-
-        const displayText = `${f.fighter_name} - ${f.fighter_type}${f.xp !== undefined ? ` (${f.xp} XP)` : ''}`;
-
-        return {
-          value: f.id,
-          displayValue: displayText,
-          label: (
-            <span className="flex items-center gap-1">
-              <span>{displayText}</span>
-              {statusIcons.length > 0 && <span className="flex items-center gap-0.5">{statusIcons}</span>}
-            </span>
-          ),
-        };
-      });
-  }, [crewFighters, positioning]);
+  const crewFighterOptions = useCrewFighterOptions(fighters, positioning);
 
   // Get all vehicles, including those assigned to fighters
   const allVehicles = useMemo<CombinedVehicleProps[]>(() => {
